pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "contract-ai-agent"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-bigquery",
    "typing_extensions",
    "streamlit",
]

[tool.setuptools.packages.find]
where = ["."]